Rename parts.ts to directive-helpers.ts
