+++ packages/lit-html/.gitignore
@@ -4,7 +4,7 @@
 /test/
 /hydrate.*
 /lit-html.*
-/parts.*
+/directive-helpers.*
 /disconnectable-directive.*
 /platform-support.*
 /private-ssr-support.*
