+++ packages/lit-html/package.json
@@ -13,9 +13,9 @@
       "development": "./development/lit-html.js",
       "default": "./lit-html.js"
     },
-    "./parts.js": {
-      "development": "./development/parts.js",
-      "default": "./parts.js"
+    "./directive-helpers.js": {
+      "development": "./development/directive-helpers.js",
+      "default": "./directive-helpers.js"
     },
     "./platform-support.js": {
       "development": "./development/platform-support.js",
@@ -41,7 +41,7 @@
     "build:ts:watch": "tsc --watch",
     "check-version": "node scripts/check-version-tracker.js",
     "checksize": "rollup -c --environment=CHECKSIZE",
-    "clean": "rm -rf lit-html.{js,js.map,d.ts} platform-support.{js,js.map,d.ts} parts.{js,js.map,d.ts} hydrate.{js,js.map,d.ts} private-ssr-support.{js,js.map,d.ts} directives/ development/",
+    "clean": "rm -rf lit-html.{js,js.map,d.ts} platform-support.{js,js.map,d.ts} directive-helpers.{js,js.map,d.ts} hydrate.{js,js.map,d.ts} private-ssr-support.{js,js.map,d.ts} directives/ development/",
     "dev": "scripts/dev.sh",
     "prepublishOnly": "npm run check-version",
     "test": "npm run test:dev && npm run test:prod",
@@ -57,10 +57,10 @@
     "/lit-html.d.ts.map",
     "/lit-html.js",
     "/lit-html.js.map",
-    "/parts.d.ts",
-    "/parts.d.ts.map",
-    "/parts.js",
-    "/parts.js.map",
+    "/directive-helpers.d.ts",
+    "/directive-helpers.d.ts.map",
+    "/directive-helpers.js",
+    "/directive-helpers.js.map",
     "/platform-support.d.ts",
     "/platform-support.d.ts.map",
     "/platform-support.js",
