+++ packages/lit-html/src/directive-helpers.ts
