+++ packages/lit-html/src/directives/repeat.ts
@@ -27,7 +27,7 @@ import {
   removePart,
   resetPartValue,
   setPartValue,
-} from '../parts.js';
+} from '../directive-helpers.js';
 
 export type KeyFn<T> = (item: T, index: number) => unknown;
 export type ItemTemplate<T> = (item: T, index: number) => unknown;
